extract base_repo + mock ref validation
