+++ src/main/java/eu/fasten/vulnerabilityproducer/utils/PatchFinder.java
@@ -22,7 +22,6 @@ import com.mongodb.client.MongoDatabase;
 import eu.fasten.vulnerabilityproducer.db.NitriteController;
 import eu.fasten.vulnerabilityproducer.db.PatchObject;
 import eu.fasten.vulnerabilityproducer.utils.connections.JavaHttpClient;
-import eu.fasten.vulnerabilityproducer.utils.connections.MongoConnector;
 import eu.fasten.vulnerabilityproducer.utils.patches.*;
 import org.jooq.tools.json.JSONParser;
 import org.slf4j.Logger;
@@ -33,6 +32,7 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.UnaryOperator;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -59,6 +59,17 @@ public class PatchFinder {
     public static final String owner = "[a-zA-Z_0-9\\-]++";
     public static final String project = "[a-zA-Z_0-9\\-\\.]++";
     public static final String sha = "[a-zA-Z_0-9/]++";
+
+    // base repo cutter
+    UnaryOperator<String> repoCutter = ref -> {
+        var sb = new StringBuilder();
+        var repo = ref.split("/");
+        if (repo.length > 4) {
+            return sb.append("https://").append(repo[2] + "/").append(repo[3] + "/").append(repo[4]).toString();
+        } else {
+            return null;
+        }
+    };
 
     /**
      * Takes a MongoDatabase object to connect and retrieve collections from.
@@ -139,7 +150,7 @@ public class PatchFinder {
                     po.setFilesChanged(vp);
                     nc.insertPatch(po);
                     logger.info("Found " + vp.size() + " patched files in ref: " + ref +
-                            "for vulnerability " + vulnerability.getId());
+                            " for vulnerability " + vulnerability.getId());
 
                     // Adding patches to the vulnerability
                     vp.forEach(p -> {
@@ -160,7 +171,7 @@ public class PatchFinder {
      * @return - set of patches extracted from the link.
      */
     public static HashSet<Vulnerability.Patch> parseReference(String ref) {
-//        ref = validateAndRedirect(ref);
+        ref = httpClient.validateAndRedirect(ref, refSet);
         if (ref == null)    return null;
         // Prevents circular links
         if (refSet.contains(ref))   {
@@ -252,33 +263,6 @@ public class PatchFinder {
     }
 
     /**
-     * Validates whether the url is valid and looks for redirects.
-     * @param url - ref to check
-     * @return ref to parse - null if ref is not valid
-     */
-    private static String validateAndRedirect(String url) {
-        try {
-            if (url == null)    return null;
-            var con = (HttpURLConnection) new URL(url).openConnection();
-            // Set timeouts to connect and read
-            con.setConnectTimeout(3000);
-            con.setReadTimeout(3000);
-            con.setInstanceFollowRedirects(false);
-            con.connect();
-            con.getInputStream();
-
-            if (con.getResponseCode() == HttpURLConnection.HTTP_MOVED_PERM || con.getResponseCode() == HttpURLConnection.HTTP_MOVED_TEMP) {
-                String redirectUrl = con.getHeaderField("Location");
-                // Prevent circular redirects
-                return refSet.contains(redirectUrl) ? null : validateAndRedirect(redirectUrl);
-            }
-            return url;
-        } catch (IOException e) {
-            return null;
-        }
-    }
-
-    /**
      * Looks in the patch links for a "repo base" url.
      * For now handle GitHub links only, since they represent the majority of the base links.
      * @param v _ Vulnerability Object
@@ -286,22 +270,10 @@ public class PatchFinder {
      */
     public String getBaseRepo(Vulnerability v) {
         HashMap<String, Integer> mapCounts = new HashMap<>();
-        v.getPatchLinks().stream().filter(ref -> ref.startsWith("https://github.com")).forEach(
-                ref -> mapCounts.merge(getBaseGitHub(ref), 1, Integer::sum)
+        v.getPatchLinks().stream().forEach(
+                ref -> mapCounts.merge((repoCutter.apply(ref)), 1, Integer::sum)
         );
         if (mapCounts.isEmpty())    return null;
-        var mostFreqBase = Collections.max(mapCounts.entrySet(), Map.Entry.comparingByValue()).getKey();
-        return mostFreqBase;
-    }
-
-    /**
-     * Extracts base url of the GitHub repo.
-     * @param ref - link to commit/pr/issue
-     * @return - base url of the repository
-     */
-    private String getBaseGitHub(String ref) {
-        assert ref.contains("github.com");
-        var repoInfo = ref.split("/");
-        return "https://github.com/" + repoInfo[3] + "/" + repoInfo[4];
+        return Collections.max(mapCounts.entrySet(), Map.Entry.comparingByValue()).getKey();
     }
 }
