+++ src/main/java/eu/fasten/vulnerabilityproducer/utils/connections/JavaHttpClient.java
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
@@ -32,6 +33,7 @@ import java.net.http.HttpResponse;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.stream.Collectors;
 
 public class JavaHttpClient {
@@ -98,5 +100,33 @@ public class JavaHttpClient {
         HttpResponse<String> response = client.send(request,
                 HttpResponse.BodyHandlers.ofString());
         return response.body();
+    }
+
+    /**
+     * Validates whether the url is valid and looks for redirects.
+     * @param url - ref to check
+     * @param refSet - set of references already seen
+     * @return ref to parse - null if ref is not valid
+     */
+    public String validateAndRedirect(String url, HashSet<String> refSet) {
+        try {
+            if (url == null)    return null;
+            var con = (HttpURLConnection) new URL(url).openConnection();
+            // Set timeouts to connect and read
+            con.setConnectTimeout(3000);
+            con.setReadTimeout(3000);
+            con.setInstanceFollowRedirects(false);
+            con.connect();
+            con.getInputStream();
+
+            if (con.getResponseCode() == HttpURLConnection.HTTP_MOVED_PERM || con.getResponseCode() == HttpURLConnection.HTTP_MOVED_TEMP) {
+                String redirectUrl = con.getHeaderField("Location");
+                // Prevent circular redirects
+                return refSet.contains(redirectUrl) ? null : validateAndRedirect(redirectUrl, refSet);
+            }
+            return url;
+        } catch (IOException e) {
+            return null;
+        }
     }
 }
