+++ src/test/java/eu/fasten/vulnerabilityproducer/PatchFinderTest.java
@@ -43,6 +43,7 @@ import java.util.stream.Collectors;
 import static com.mongodb.client.model.Filters.eq;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 
 public class PatchFinderTest {
@@ -166,6 +167,9 @@ public class PatchFinderTest {
         when(dbMock.getCollection("pull_requests")).thenReturn(prsMock);
         when(dbMock.getCollection("issues")).thenReturn(issuesMock);
         patchFinder = new PatchFinder(dbMock, clientMock);
+
+        // Do not validate refs when testing
+        when(clientMock.validateAndRedirect(Mockito.anyString(), Mockito.any())).thenAnswer(i -> i.getArguments()[0]);
     }
 
     public static String readFile(String filePath) throws IOException {
